More minor UI modifications and improvements
