--- Example App/lib/src/Core/Interactions/SensationActionsList.dart
@@ -1,1 +1,0 @@
-part of AureusTestingApp;
