+++ Example App/lib/src/Core/Interactions/Sensations/SensationActionsList.dart
@@ -1,0 +1,1 @@
+part of AureusTestingApp;
