+++ Example App/lib/src/Core/Landing/HomeNavigation.dart
@@ -13,28 +13,35 @@ class _ExplorationViewState extends State<ExplorationView> {
         accessibilityHint: "Shows you all of the UI Components in Aureus.",
         tabIcon: Assets.add),
     ControllerTabObject(
-        tabController: AureusInteractionsView(),
-        tabTitle: "Interact",
-        accessibilityHint:
-            "Lets you try the interactibility aspects of Aureus.",
-        tabIcon: Assets.brain),
-    ControllerTabObject(
         tabController: ToolDetailView(
           parentTool: testingTool,
         ),
         tabTitle: "Tools",
         accessibilityHint: "Shows you the tools",
         tabIcon: Assets.settings),
+    /*
     ControllerTabObject(
         tabController: AureusSafetyView(),
         tabTitle: "Safety",
         accessibilityHint: "Shows you the safety features within Aurues",
         tabIcon: Assets.person),
+        ControllerTabObject(
+        tabController: AureusInteractionsView(),
+        tabTitle: "Interact",
+        accessibilityHint:
+            "Lets you try the interactibility aspects of Aureus.",
+        tabIcon: Assets.brain),
+    */
   ];
 
   @override
   Widget build(BuildContext context) {
-    return VerticalSideNavBarComponent(tabItems: tabItems);
+    return VerticalSideNavBarComponent(
+      tabItems: tabItems,
+      altColor: palette.brightness() == Brightness.light
+          ? palette.white()
+          : palette.black(),
+    );
   }
 }
