+++ Example App/lib/src/Core/UI Components/Components/ComponentsSubView.dart
@@ -45,7 +45,7 @@ class _ComponentsSubviewState extends State<ComponentsSubview> {
     ], containerVariant: wrapperVariants.stackScroll);
 
     return ContainerView(
-      decorationVariant: decorationPriority.standard,
+      decorationVariant: decorationPriority.important,
       builder: viewBuilder,
       showQuickActionBar: false,
     );
