+++ Example App/lib/src/Core/UI Components/Elements/ElementsSubView.dart
@@ -45,7 +45,7 @@ class _ElementsSubviewState extends State<ElementsSubview> {
     ], containerVariant: wrapperVariants.stackScroll);
 
     return ContainerView(
-      decorationVariant: decorationPriority.standard,
+      decorationVariant: decorationPriority.important,
       builder: viewBuilder,
       showQuickActionBar: false,
     );
