+++ Package/lib/src/Widgets/Components/Tabbing Bars/NavBarComponent.dart
@@ -44,7 +44,7 @@ class _NavBarComponentState extends State<NavBarComponent> {
 
       var tabItem = BottomNavigationBarItem(
         icon: Icon(element.tabIcon, size: size.responsiveSize(38.0)),
-        label: ".",
+        label: element.accessibilityHint,
         tooltip: element.accessibilityHint,
         backgroundColor: Colors.transparent,
       );
@@ -58,7 +58,7 @@ class _NavBarComponentState extends State<NavBarComponent> {
       elevation: 0,
       selectedItemColor: coloration.contrastColor(),
       unselectedItemColor: coloration.inactiveColor(),
-      showSelectedLabels: true,
+      showSelectedLabels: false,
       showUnselectedLabels: false,
       backgroundColor: Colors.transparent,
       items: tabbingArray,
