+++ Package/lib/src/Widgets/Elements/Cards/Standard/StandardCardElement.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: must_be_immutable
+
 import 'package:aureus/aureus.dart';
 
 /// @nodoc
