+++ Package/lib/src/Widgets/Elements/Info Hierarchy/ArticleViewElement.dart
@@ -26,30 +26,21 @@ class ArticleViewElement extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var screenSize = size.logicalScreenSize();
-
-    return SizedBox(
-      width: size.layoutItemWidth(1, screenSize),
-      height: size.layoutItemHeight(1, screenSize),
-      child: SingleChildScrollView(
-        scrollDirection: Axis.vertical,
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisAlignment: MainAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const SizedBox(height: 10.0),
-            HeadingOneText(title, decorationPriority.standard),
-            const SizedBox(height: 10.0),
-            SubheaderText(subheader, decorationPriority.standard),
-            const SizedBox(height: 10.0),
-            const DividerElement(),
-            const SizedBox(height: 10.0),
-            BodyOneText(body, decorationPriority.standard),
-            const SizedBox(height: 10.0),
-          ],
-        ),
-      ),
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      mainAxisAlignment: MainAxisAlignment.start,
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        const SizedBox(height: 10.0),
+        HeadingOneText(title, decorationPriority.standard),
+        const SizedBox(height: 10.0),
+        SubheaderText(subheader, decorationPriority.standard),
+        const SizedBox(height: 10.0),
+        const DividerElement(),
+        const SizedBox(height: 10.0),
+        BodyOneText(body, decorationPriority.standard),
+        const SizedBox(height: 10.0),
+      ],
     );
   }
 }
